refactor(scripts): split analytics-core helper clusters into lib/analytics/
Move-only extraction of the 39 private helper functions from
lib/analytics-core.ts into domain modules under lib/analytics/:
types.ts (shared interfaces/type aliases), snapshots.ts,
listing-meta.ts, map-statistics.ts, by-day.ts. Function bodies,
types, and constants are moved verbatim; runGenerateAnalyticsCli
and its CLI helpers (validateArgs, parseTopK, limitRows) plus the
megafunction-only row interfaces stay in analytics-core.ts. The
public surface (runGenerateAnalyticsCli) is unchanged.
