+++ scripts/lib/analytics-core.ts
@@ -1,48 +1,45 @@
-import { existsSync, mkdirSync, rmSync } from "node:fs";
-import { join, resolve } from "node:path";
+import { mkdirSync, rmSync } from "node:fs";
+import { join } from "node:path";
 import { pathToFileURL } from "node:url";
+import { loadAuthorAliasIndex } from "./author-aliases.js";
 import { getFlagValue } from "./cli.js";
 import { writeCsv } from "./csv.js";
-import { findLastAtOrBefore, listSnapshotFileNames } from "./history-utils.js";
-import { isFiniteNumber, isObject, readJsonFile } from "./json-utils.js";
+import { readJsonFile } from "./json-utils.js";
 import { resolveRepoRoot } from "./script-runtime.js";
-import { isTestListing } from "./test-listings.js";
-import { loadAuthorAliasIndex, resolveAuthorPresentation, type AuthorAliasIndex } from "./author-aliases.js";
-
-interface SnapshotEntry {
-  file: string;
-  date: Date;
-}
-
-interface SnapshotData {
-  schema_version?: unknown;
-  snapshot_date?: unknown;
-  generated_at?: unknown;
-  total_downloads?: unknown;
-  raw_total_downloads?: unknown;
-  total_attributed_downloads?: unknown;
-  net_downloads?: unknown;
-  maps?: { downloads?: Record<string, Record<string, unknown>> };
-  mods?: { downloads?: Record<string, Record<string, unknown>> };
-}
-
-type ListingTotals = Map<ListingKey, number>;
-
-interface ListingMeta {
-  name: string;
-  author: string;
-  author_alias: string;
-  attribution_link: string;
-  github_id: number | null;
-}
-
-interface ListingProjectRow {
-  listing_type: "map" | "mod";
-  id: string;
-  name: string;
-  project_key: string;
-  project_name: string;
-}
+import {
+  buildAssetsByDayRows,
+  buildAuthorByDayRows,
+  buildDailyDeltaSnapshotTotals,
+  buildListingByDayRows,
+  buildProjectByDayRows,
+  buildSignedDailyDeltaSnapshotTotals,
+  toSnapshotDateLabel,
+} from "./analytics/by-day.js";
+import {
+  filterOutTestListingTotals,
+  loadListingProjectRow,
+  loadManifestMeta,
+  toListingLabel,
+} from "./analytics/listing-meta.js";
+import { loadMapStatisticsRows } from "./analytics/map-statistics.js";
+import {
+  buildListingIdsBySnapshot,
+  buildListingVersionsBySnapshot,
+  buildMonotonicSnapshotTotals,
+  listSnapshots,
+  resolveBaselineSnapshot,
+  toListingTotals,
+} from "./analytics/snapshots.js";
+import type {
+  AssetByDayRow,
+  DailySeriesRow,
+  ListingKey,
+  ListingMeta,
+  ListingProjectRow,
+  ListingTotals,
+  MapStatisticsRow,
+  SnapshotData,
+} from "./analytics/types.js";
 
 interface ListingWindowRow {
   rank: number;
@@ -148,70 +145,6 @@ interface AuthorWindowRow {
   baseline_snapshot: string;
 }
 
-interface MapStatisticsRow {
-  rank: number;
-  id: string;
-  name: string;
-  author: string;
-  author_alias: string;
-  attribution_link: string;
-  city_code: string;
-  country: string;
-  population: number;
-  population_count: number;
-  points_count: number;
-  n_cells: number;
-  mean_point_density: number;
-  median_resident_weighted_nn_km: number;
-  mean_resident_weighted_nn_km: number;
-  median_worker_weighted_nn_km: number;
-  mean_worker_weighted_nn_km: number;
-  detail_radius_km: number;
-  detail_score: number;
-  playable_area_cells: number;
-  median_cell_resident_density: number;
-  mean_cell_resident_density: number;
-  pct_cells_with_residents: number;
-  median_cell_worker_density: number;
-  mean_cell_worker_density: number;
-  pct_cells_with_workers: number;
-  median_commute_distance: number;
-  mean_commute_distance: number;
-  detected_center_count: number;
-  polycentrism_score: number;
-}
-
-interface AssetByDayRow {
-  snapshot_date: string;
-  total_downloads: number;
-  maps: number;
-  mods: number;
-  total_downloads_signed: number;
-  maps_signed: number;
-  mods_signed: number;
-  total_downloads_clamped: number;
-  maps_clamped: number;
-  mods_clamped: number;
-  cumulative_total: number;
-  cumulative_maps: number;
-  cumulative_mods: number;
-  total_new_assets: number;
-  new_maps: number;
-  new_mods: number;
-  cumulative_assets: number;
-  cumulative_maps_assets: number;
-  cumulative_mods_assets: number;
-  total_new_assets_versions: number;
-  new_maps_versions: number;
-  new_mods_versions: number;
-  cumulative_asset_versions: number;
-  cumulative_maps_versions: number;
-  cumulative_mods_versions: number;
-}
-
-type DailySeriesRow = Record<string, string | number>;
-type ListingKey = `${"maps" | "mods"}:${string}`;
-
 const DEFAULT_TOP_LISTINGS: number | null = null;
 const DEFAULT_TOP_AUTHORS: number | null = null;
 const WINDOWS = [1, 3, 7, 14, 30] as const;
@@ -246,919 +179,6 @@ function parseTopK(rawValue: string | undefined, fallback: number | null, label:
 
 function limitRows<T>(rows: T[], topK: number | null): T[] {
   return topK === null ? rows : rows.slice(0, topK);
-}
-
-function parseSnapshotDate(fileName: string): Date | null {
-  const match = fileName.match(/^snapshot_(\d{4})_(\d{2})_(\d{2})\.json$/);
-  if (!match) return null;
-  const date = new Date(Date.UTC(Number(match[1]), Number(match[2]) - 1, Number(match[3])));
-  return Number.isNaN(date.getTime()) ? null : date;
-}
-
-function listSnapshots(historyDir: string): SnapshotEntry[] {
-  return listSnapshotFileNames(historyDir)
-    .map((file) => ({ file, date: parseSnapshotDate(file) }))
-    .filter((entry): entry is SnapshotEntry => entry.date instanceof Date)
-    .sort((a, b) => a.date.getTime() - b.date.getTime());
-}
-
-function toListingTotals(snapshot: SnapshotData): ListingTotals {
-  const totals = new Map<ListingKey, number>();
-  for (const listingType of ["maps", "mods"] as const) {
-    // `downloads` is the canonical adjusted/non-attributed listing data in schema v2.
-    const downloads = snapshot?.[listingType]?.downloads;
-    if (!downloads || typeof downloads !== "object") continue;
-    for (const [id, versions] of Object.entries(downloads)) {
-      if (!versions || typeof versions !== "object") continue;
-      let total = 0;
-      for (const count of Object.values(versions)) {
-        if (isFiniteNumber(count)) {
-          total += count;
-        }
-      }
-      totals.set(`${listingType}:${id}`, total);
-    }
-  }
-  return totals;
-}
-
-function toListingIdSet(
-  snapshot: SnapshotData,
-  repoRoot: string,
-  listingType: "maps" | "mods",
-): Set<string> {
-  const ids = new Set<string>();
-  const downloads = snapshot?.[listingType]?.downloads;
-  if (!downloads || typeof downloads !== "object") return ids;
-  for (const id of Object.keys(downloads)) {
-    if (isTestListing(repoRoot, listingType, id)) continue;
-    ids.add(id);
-  }
-  return ids;
-}
-
-function buildListingIdsBySnapshot(
-  snapshots: SnapshotEntry[],
-  historyDir: string,
-  repoRoot: string,
-): Map<string, { maps: Set<string>; mods: Set<string> }> {
-  const bySnapshot = new Map<string, { maps: Set<string>; mods: Set<string> }>();
-  for (const snapshot of snapshots) {
-    const snapshotData = readJsonFile<SnapshotData>(join(historyDir, snapshot.file));
-    bySnapshot.set(snapshot.file, {
-      maps: toListingIdSet(snapshotData, repoRoot, "maps"),
-      mods: toListingIdSet(snapshotData, repoRoot, "mods"),
-    });
-  }
-  return bySnapshot;
-}
-
-function toListingVersionSet(
-  snapshot: SnapshotData,
-  repoRoot: string,
-  listingType: "maps" | "mods",
-): Set<string> {
-  const versions = new Set<string>();
-  const downloads = snapshot?.[listingType]?.downloads;
-  if (!downloads || typeof downloads !== "object") return versions;
-
-  for (const [id, versionMap] of Object.entries(downloads)) {
-    if (isTestListing(repoRoot, listingType, id)) continue;
-    if (!versionMap || typeof versionMap !== "object") continue;
-    for (const version of Object.keys(versionMap)) {
-      versions.add(`${id}@@${version}`);
-    }
-  }
-
-  return versions;
-}
-
-function buildListingVersionsBySnapshot(
-  snapshots: SnapshotEntry[],
-  historyDir: string,
-  repoRoot: string,
-): Map<string, { maps: Set<string>; mods: Set<string> }> {
-  const bySnapshot = new Map<string, { maps: Set<string>; mods: Set<string> }>();
-  for (const snapshot of snapshots) {
-    const snapshotData = readJsonFile<SnapshotData>(join(historyDir, snapshot.file));
-    bySnapshot.set(snapshot.file, {
-      maps: toListingVersionSet(snapshotData, repoRoot, "maps"),
-      mods: toListingVersionSet(snapshotData, repoRoot, "mods"),
-    });
-  }
-  return bySnapshot;
-}
-
-function buildMonotonicSnapshotTotals(
-  snapshots: SnapshotEntry[],
-  historyDir: string,
-): Map<string, ListingTotals> {
-  const monotonicBySnapshot = new Map<string, ListingTotals>();
-  const runningMax = new Map<ListingKey, number>();
-
-  for (const snapshot of snapshots) {
-    const snapshotData = readJsonFile<SnapshotData>(join(historyDir, snapshot.file));
-    const adjustedTotals = toListingTotals(snapshotData);
-    const snapshotMonotonic = new Map<ListingKey, number>();
-    const keys = new Set<ListingKey>([
-      ...runningMax.keys(),
-      ...adjustedTotals.keys(),
-    ]);
-
-    for (const key of keys) {
-      const previousMax = runningMax.get(key) ?? 0;
-      const adjustedTotal = adjustedTotals.get(key) ?? 0;
-      const nextMax = Math.max(previousMax, adjustedTotal);
-      if (nextMax > 0 || adjustedTotals.has(key) || runningMax.has(key)) {
-        runningMax.set(key, nextMax);
-        snapshotMonotonic.set(key, nextMax);
-      }
-    }
-
-    monotonicBySnapshot.set(snapshot.file, snapshotMonotonic);
-  }
-
-  return monotonicBySnapshot;
-}
-
-function addDays(date: Date, days: number): Date {
-  const copy = new Date(date.getTime());
-  copy.setUTCDate(copy.getUTCDate() + days);
-  return copy;
-}
-
-function findSnapshotAtOrBefore(snapshots: SnapshotEntry[], targetDate: Date): SnapshotEntry | null {
-  return findLastAtOrBefore(snapshots, (entry) => entry.date.getTime(), targetDate.getTime());
-}
-
-function resolveBaselineSnapshot(snapshots: SnapshotEntry[], latestDate: Date, days: number): SnapshotEntry {
-  const target = addDays(latestDate, -days);
-  return findSnapshotAtOrBefore(snapshots, target) ?? snapshots[0];
-}
-
-function toListingLabel(listingType: "maps" | "mods"): "map" | "mod" {
-  return listingType === "maps" ? "map" : "mod";
-}
-
-function filterOutTestListingTotals(
-  repoRoot: string,
-  totals: ListingTotals,
-): ListingTotals {
-  const filtered = new Map<ListingKey, number>();
-  for (const [key, total] of totals.entries()) {
-    const [listingType, id] = key.split(":") as ["maps" | "mods", string];
-    if (isTestListing(repoRoot, listingType, id)) continue;
-    filtered.set(key, total);
-  }
-  return filtered;
-}
-
-function loadManifestMeta(
-  repoRoot: string,
-  listingType: "maps" | "mods",
-  id: string,
-  authorAliases: AuthorAliasIndex,
-): ListingMeta {
-  const manifestPath = join(repoRoot, listingType, id, "manifest.json");
-  try {
-    const manifest = readJsonFile<Record<string, unknown>>(manifestPath);
-    const name = typeof manifest.name === "string" && manifest.name.trim() !== ""
-      ? manifest.name
-      : id;
-    const author = typeof manifest.author === "string" && manifest.author.trim() !== ""
-      ? manifest.author
-      : "UNKNOWN";
-    const githubId = typeof manifest.github_id === "number" && Number.isFinite(manifest.github_id)
-      ? manifest.github_id
-      : null;
-    const presentation = resolveAuthorPresentation(author, githubId, authorAliases);
-    return {
-      name,
-      author: presentation.author,
-      author_alias: presentation.author_alias,
-      attribution_link: presentation.attribution_link,
-      github_id: githubId,
-    };
-  } catch {
-    return {
-      name: id,
-      author: "UNKNOWN",
-      author_alias: "UNKNOWN",
-      attribution_link: "https://github.com/UNKNOWN",
-      github_id: null,
-    };
-  }
-}
-
-function parseProjectFromUrl(urlValue: string): { projectKey: string; projectName: string } | null {
-  try {
-    const parsed = new URL(urlValue);
-    const segments = parsed.pathname.split("/").filter((segment) => segment !== "");
-
-    if (parsed.hostname === "github.com" && segments.length >= 2) {
-      const owner = segments[0]!;
-      const repo = segments[1]!;
-      return {
-        projectKey: `${owner.toLowerCase()}/${repo.toLowerCase()}`,
-        projectName: repo,
-      };
-    }
-
-    if (parsed.hostname === "raw.githubusercontent.com" && segments.length >= 2) {
-      const owner = segments[0]!;
-      const repo = segments[1]!;
-      return {
-        projectKey: `${owner.toLowerCase()}/${repo.toLowerCase()}`,
-        projectName: repo,
-      };
-    }
-
-    if (parsed.hostname.endsWith(".github.io")) {
-      const owner = parsed.hostname.slice(0, -".github.io".length);
-      const repo = segments[0] ?? owner;
-      return {
-        projectKey: `${owner.toLowerCase()}/${repo.toLowerCase()}`,
-        projectName: repo,
-      };
-    }
-  } catch {
-    return null;
-  }
-
-  return null;
-}
-
-function loadListingProjectRow(
-  repoRoot: string,
-  listingType: "maps" | "mods",
-  id: string,
-): ListingProjectRow {
-  const manifestPath = join(repoRoot, listingType, id, "manifest.json");
-  const listingLabel = toListingLabel(listingType);
-
-  try {
-    const manifest = readJsonFile<Record<string, unknown>>(manifestPath);
-    const name = typeof manifest.name === "string" && manifest.name.trim() !== ""
-      ? manifest.name
-      : id;
-
-    const sourceUrl = typeof manifest.source === "string" ? manifest.source.trim() : "";
-    const update = isObject(manifest.update) ? manifest.update : null;
-    const updateRepo = typeof update?.repo === "string" ? update.repo.trim() : "";
-    const updateUrl = typeof update?.url === "string" ? update.url.trim() : "";
-
-    const parsedProject = (
-      (sourceUrl !== "" ? parseProjectFromUrl(sourceUrl) : null)
-      ?? (updateUrl !== "" ? parseProjectFromUrl(updateUrl) : null)
-      ?? (updateRepo !== ""
-        ? {
-          projectKey: updateRepo.toLowerCase(),
-          projectName: updateRepo.split("/").pop() ?? updateRepo,
-        }
-        : null)
-    );
-
-    return {
-      listing_type: listingLabel,
-      id,
-      name,
-      project_key: parsedProject?.projectKey ?? `${listingLabel}:${id}`,
-      project_name: parsedProject?.projectName ?? name,
-    };
-  } catch {
-    return {
-      listing_type: listingLabel,
-      id,
-      name: id,
-      project_key: `${listingLabel}:${id}`,
-      project_name: id,
-    };
-  }
-}
-
-function toNonEmptyString(value: unknown, fallback: string): string {
-  return typeof value === "string" && value.trim() !== "" ? value : fallback;
-}
-
-function toNonNegativeNumber(value: unknown): number {
-  return typeof value === "number" && Number.isFinite(value) && value >= 0 ? value : 0;
-}
-
-function medianOf(values: number[]): number {
-  if (values.length === 0) return 0;
-  const sorted = [...values].sort((a, b) => a - b);
-  return sorted[Math.floor(sorted.length / 2)] ?? 0;
-}
-
-interface GridCellFeatureProperties {
-  jobs?: unknown;
-  pop?: unknown;
-  pointCount?: unknown;
-}
-
-interface GridMetricBundleProperties {
-  p10?: unknown;
-  p25?: unknown;
-  p50?: unknown;
-  p75?: unknown;
-  p90?: unknown;
-  mean?: unknown;
-}
-
-interface GridPolycentrismActivityProperties {
-  continuousScore?: unknown;
-  detectedCenterCount?: unknown;
-}
-
-interface GridPolycentrismProperties {
-  activity?: unknown;
-}
-
-interface GridDetailProperties {
-  radiusKm?: unknown;
-  playableAreaKm2?: unknown;
-  playableAreaPerPointKm2?: unknown;
-  playableCatchmentRadiusKm?: unknown;
-  score?: unknown;
-}
-
-interface GridSummary {
-  n_cells: number;
-  mean_point_density: number;
-  median_resident_weighted_nn_km: number;
-  mean_resident_weighted_nn_km: number;
-  median_worker_weighted_nn_km: number;
-  mean_worker_weighted_nn_km: number;
-  detail_radius_km: number;
-  detail_score: number;
-  playable_area_cells: number;
-  median_cell_resident_density: number;
-  mean_cell_resident_density: number;
-  pct_cells_with_residents: number;
-  median_cell_worker_density: number;
-  mean_cell_worker_density: number;
-  pct_cells_with_workers: number;
-  median_commute_distance: number;
-  mean_commute_distance: number;
-  detected_center_count: number;
-  polycentrism_score: number;
-}
-
-function emptyGridSummary(): GridSummary {
-  return {
-    n_cells: 0,
-    mean_point_density: 0,
-    median_resident_weighted_nn_km: 0,
-    mean_resident_weighted_nn_km: 0,
-    median_worker_weighted_nn_km: 0,
-    mean_worker_weighted_nn_km: 0,
-    detail_radius_km: 0,
-    detail_score: 0,
-    playable_area_cells: 0,
-    median_cell_resident_density: 0,
-    mean_cell_resident_density: 0,
-    pct_cells_with_residents: 0,
-    median_cell_worker_density: 0,
-    mean_cell_worker_density: 0,
-    pct_cells_with_workers: 0,
-    median_commute_distance: 0,
-    mean_commute_distance: 0,
-    detected_center_count: 0,
-    polycentrism_score: 0,
-  };
-}
-
-function roundTo(value: number, decimalPlaces = 2): number {
-  const factor = 10 ** decimalPlaces;
-  return Math.round(value * factor) / factor;
-}
-
-function nonZeroMedian(values: number[]): number {
-  return medianOf(values.filter((value) => value > 0));
-}
-
-function nonZeroMean(values: number[]): number {
-  const filtered = values.filter((value) => value > 0);
-  if (filtered.length === 0) return 0;
-  return filtered.reduce((sum, value) => sum + value, 0) / filtered.length;
-}
-
-function percentNonZero(values: number[], totalCount: number): number {
-  if (totalCount <= 0) return 0;
-  return (values.filter((value) => value > 0).length / totalCount) * 100;
-}
-
-function readGridMetricBundle(
-  properties: Record<string, unknown>,
-  key: string,
-): { p50: number; mean: number } {
-  const bundle = isObject(properties[key]) ? properties[key] as GridMetricBundleProperties : null;
-  if (!bundle) {
-    return { p50: 0, mean: 0 };
-  }
-  return {
-    p50: toNonNegativeNumber(bundle.p50),
-    mean: toNonNegativeNumber(bundle.mean),
-  };
-}
-
-function readGridDetailSummary(
-  properties: Record<string, unknown>,
-): {
-  present: boolean;
-  radiusKm: number;
-  playableAreaKm2: number;
-  playableAreaPerPointKm2: number;
-  playableCatchmentRadiusKm: number;
-  score: number;
-} {
-  const detail = isObject(properties.detail)
-    ? properties.detail as GridDetailProperties
-    : null;
-
-  if (!detail) {
-    return {
-      present: false,
-      radiusKm: 0,
-      playableAreaKm2: 0,
-      playableAreaPerPointKm2: 0,
-      playableCatchmentRadiusKm: 0,
-      score: 0,
-    };
-  }
-
-  const playableAreaKm2 = toNonNegativeNumber(detail.playableAreaKm2);
-  const playableAreaPerPointKm2 = toNonNegativeNumber(detail.playableAreaPerPointKm2);
-  const playableCatchmentRadiusKm = toNonNegativeNumber(detail.playableCatchmentRadiusKm);
-  const hasPlayableAreaShape = (
-    playableAreaKm2 > 0
-    || playableAreaPerPointKm2 > 0
-    || playableCatchmentRadiusKm > 0
-  );
-
-  return {
-    present: hasPlayableAreaShape,
-    radiusKm: toNonNegativeNumber(detail.radiusKm),
-    playableAreaKm2,
-    playableAreaPerPointKm2,
-    playableCatchmentRadiusKm,
-    score: toNonNegativeNumber(detail.score),
-  };
-}
-
-function readGridPolycentrismSummary(
-  properties: Record<string, unknown>,
-): { detectedCenterCount: number; continuousScore: number } {
-  const polycentrism = isObject(properties.polycentrism)
-    ? properties.polycentrism as GridPolycentrismProperties
-    : null;
-  const activity = polycentrism && isObject(polycentrism.activity)
-    ? polycentrism.activity as GridPolycentrismActivityProperties
-    : null;
-
-  return {
-    detectedCenterCount: toNonNegativeNumber(activity?.detectedCenterCount),
-    continuousScore: toNonNegativeNumber(activity?.continuousScore),
-  };
-}
-
-function loadGridSummary(repoRoot: string, id: string): GridSummary {
-  const gridPath = join(repoRoot, "maps", id, "grid.geojson");
-  if (!existsSync(gridPath)) return emptyGridSummary();
-
-  try {
-    const grid = readJsonFile<Record<string, unknown>>(gridPath);
-    const features = Array.isArray(grid.features) ? grid.features : [];
-    const gridProperties = isObject(grid.properties) ? grid.properties : {};
-    const commuteSummary = readGridMetricBundle(gridProperties, "commuteDistanceKm");
-    const residentWeightedNearestNeighborSummary = readGridMetricBundle(gridProperties, "residentWeightedNearestNeighborKm");
-    const workerWeightedNearestNeighborSummary = readGridMetricBundle(gridProperties, "workerWeightedNearestNeighborKm");
-    const detailSummary = readGridDetailSummary(gridProperties);
-    const polycentrismSummary = readGridPolycentrismSummary(gridProperties);
-    const populatedCells = features
-      .map((feature) => {
-        if (!isObject(feature)) return null;
-        const properties = isObject(feature.properties)
-          ? feature.properties as GridCellFeatureProperties
-          : {};
-        const pointCount = toNonNegativeNumber(properties.pointCount);
-        if (pointCount <= 0) return null;
-        return {
-          pointCount,
-          pop: toNonNegativeNumber(properties.pop),
-          jobs: toNonNegativeNumber(properties.jobs),
-        };
-      })
-      .filter((feature): feature is { pointCount: number; pop: number; jobs: number } => feature !== null);
-
-    const nCells = populatedCells.length;
-    const pointCounts = populatedCells.map((cell) => cell.pointCount);
-    const residentCounts = populatedCells.map((cell) => cell.pop);
-    const workerCounts = populatedCells.map((cell) => cell.jobs);
-    const totalPoints = pointCounts.reduce((sum, value) => sum + value, 0);
-    const detailRadiusKm = detailSummary.present ? detailSummary.radiusKm : 0;
-    const detailScore = detailSummary.present ? detailSummary.score : 0;
-    const playableAreaCellCount = detailSummary.present ? detailSummary.playableAreaKm2 : 0;
-    if (nCells === 0) {
-      return {
-        ...emptyGridSummary(),
-        median_resident_weighted_nn_km: roundTo(residentWeightedNearestNeighborSummary.p50, 3),
-        mean_resident_weighted_nn_km: roundTo(residentWeightedNearestNeighborSummary.mean, 3),
-        median_worker_weighted_nn_km: roundTo(workerWeightedNearestNeighborSummary.p50, 3),
-        mean_worker_weighted_nn_km: roundTo(workerWeightedNearestNeighborSummary.mean, 3),
-        detail_radius_km: roundTo(detailRadiusKm, 3),
-        detail_score: roundTo(detailScore),
-        playable_area_cells: Math.round(playableAreaCellCount),
-        median_commute_distance: commuteSummary.p50,
-        mean_commute_distance: commuteSummary.mean,
-        detected_center_count: polycentrismSummary.detectedCenterCount,
-        polycentrism_score: roundTo(polycentrismSummary.continuousScore),
-      };
-    }
-
-    return {
-      n_cells: nCells,
-      mean_point_density: roundTo(totalPoints / nCells),
-      median_resident_weighted_nn_km: roundTo(residentWeightedNearestNeighborSummary.p50, 3),
-      mean_resident_weighted_nn_km: roundTo(residentWeightedNearestNeighborSummary.mean, 3),
-      median_worker_weighted_nn_km: roundTo(workerWeightedNearestNeighborSummary.p50, 3),
-      mean_worker_weighted_nn_km: roundTo(workerWeightedNearestNeighborSummary.mean, 3),
-      detail_radius_km: roundTo(detailRadiusKm, 3),
-      detail_score: roundTo(detailScore),
-      playable_area_cells: Math.round(playableAreaCellCount),
-      median_cell_resident_density: roundTo(nonZeroMedian(residentCounts)),
-      mean_cell_resident_density: roundTo(nonZeroMean(residentCounts)),
-      pct_cells_with_residents: roundTo(percentNonZero(residentCounts, nCells)),
-      median_cell_worker_density: roundTo(nonZeroMedian(workerCounts)),
-      mean_cell_worker_density: roundTo(nonZeroMean(workerCounts)),
-      pct_cells_with_workers: roundTo(percentNonZero(workerCounts, nCells)),
-      median_commute_distance: roundTo(commuteSummary.p50),
-      mean_commute_distance: roundTo(commuteSummary.mean),
-      detected_center_count: polycentrismSummary.detectedCenterCount,
-      polycentrism_score: roundTo(polycentrismSummary.continuousScore),
-    };
-  } catch {
-    return emptyGridSummary();
-  }
-}
-
-function loadMapStatisticsRows(repoRoot: string, authorAliases: AuthorAliasIndex): MapStatisticsRow[] {
-  const indexPath = join(repoRoot, "maps", "index.json");
-  const index = readJsonFile<{ maps?: unknown }>(indexPath);
-  const mapIds = Array.isArray(index.maps)
-    ? index.maps.filter((value): value is string => typeof value === "string")
-    : [];
-
-  const rows: Omit<MapStatisticsRow, "rank">[] = [];
-  for (const id of mapIds) {
-    const manifestPath = join(repoRoot, "maps", id, "manifest.json");
-    try {
-      const manifest = readJsonFile<Record<string, unknown>>(manifestPath);
-      const author = toNonEmptyString(manifest.author, "UNKNOWN");
-      const githubId = typeof manifest.github_id === "number" && Number.isFinite(manifest.github_id)
-        ? manifest.github_id
-        : null;
-      const presentation = resolveAuthorPresentation(author, githubId, authorAliases);
-      const gridSummary = loadGridSummary(repoRoot, id);
-      rows.push({
-        id,
-        name: toNonEmptyString(manifest.name, id),
-        author: presentation.author,
-        author_alias: presentation.author_alias,
-        attribution_link: presentation.attribution_link,
-        city_code: toNonEmptyString(manifest.city_code, ""),
-        country: toNonEmptyString(manifest.country, ""),
-        population: toNonNegativeNumber(manifest.population),
-        population_count: toNonNegativeNumber(manifest.population_count),
-        points_count: toNonNegativeNumber(manifest.points_count),
-        ...gridSummary,
-      });
-    } catch {
-      rows.push({
-        id,
-        name: id,
-        author: "UNKNOWN",
-        author_alias: "UNKNOWN",
-        attribution_link: "https://github.com/UNKNOWN",
-        city_code: "",
-        country: "",
-        population: 0,
-        population_count: 0,
-        points_count: 0,
-        ...emptyGridSummary(),
-      });
-    }
-  }
-
-  rows.sort((a, b) =>
-    b.population - a.population
-    || b.n_cells - a.n_cells
-    || b.population_count - a.population_count
-    || b.points_count - a.points_count
-    || a.id.localeCompare(b.id));
-
-  return rows.map((row, index) => ({
-    rank: index + 1,
-    ...row,
-  }));
-}
-
-function toSnapshotDateLabel(fileName: string): string {
-  const match = fileName.match(/^snapshot_(\d{4}_\d{2}_\d{2})\.json$/);
-  return match?.[1] ?? fileName;
-}
-
-function buildDailyDeltaSnapshotTotals(
-  snapshots: SnapshotEntry[],
-  monotonicTotalsBySnapshot: Map<string, ListingTotals>,
-): Map<string, ListingTotals> {
-  const bySnapshot = new Map<string, ListingTotals>();
-  let previousTotals = new Map<ListingKey, number>();
-
-  for (const snapshot of snapshots) {
-    const currentTotals = monotonicTotalsBySnapshot.get(snapshot.file) ?? new Map<ListingKey, number>();
-    const keys = new Set<ListingKey>([
-      ...previousTotals.keys(),
-      ...currentTotals.keys(),
-    ]);
-    const deltaTotals = new Map<ListingKey, number>();
-    for (const key of keys) {
-      const delta = Math.max(0, (currentTotals.get(key) ?? 0) - (previousTotals.get(key) ?? 0));
-      if (delta > 0 || currentTotals.has(key) || previousTotals.has(key)) {
-        deltaTotals.set(key, delta);
-      }
-    }
-    bySnapshot.set(snapshot.file, deltaTotals);
-    previousTotals = currentTotals;
-  }
-
-  return bySnapshot;
-}
-
-function buildSignedDailyDeltaSnapshotTotals(
-  snapshots: SnapshotEntry[],
-  adjustedTotalsBySnapshot: Map<string, ListingTotals>,
-): Map<string, ListingTotals> {
-  const bySnapshot = new Map<string, ListingTotals>();
-  let previousTotals = new Map<ListingKey, number>();
-
-  for (const snapshot of snapshots) {
-    const currentTotals = adjustedTotalsBySnapshot.get(snapshot.file) ?? new Map<ListingKey, number>();
-    const keys = new Set<ListingKey>([
-      ...previousTotals.keys(),
-      ...currentTotals.keys(),
-    ]);
-    const deltaTotals = new Map<ListingKey, number>();
-    for (const key of keys) {
-      const delta = (currentTotals.get(key) ?? 0) - (previousTotals.get(key) ?? 0);
-      if (delta !== 0 || currentTotals.has(key) || previousTotals.has(key)) {
-        deltaTotals.set(key, delta);
-      }
-    }
-    bySnapshot.set(snapshot.file, deltaTotals);
-    previousTotals = currentTotals;
-  }
-
-  return bySnapshot;
-}
-
-function buildListingByDayRows(
-  snapshotDates: string[],
-  latestTotals: ListingTotals,
-  dailyDeltasBySnapshot: Map<string, ListingTotals>,
-  listingMeta: Map<ListingKey, ListingMeta>,
-  listingProjectByKey: Map<ListingKey, ListingProjectRow>,
-): DailySeriesRow[] {
-  const rows: DailySeriesRow[] = [];
-  for (const [key, totalDownloads] of latestTotals.entries()) {
-    const [listingType, id] = key.split(":") as ["maps" | "mods", string];
-    const meta = listingMeta.get(key) ?? {
-      name: id,
-      author: "UNKNOWN",
-      author_alias: "UNKNOWN",
-      attribution_link: "https://github.com/UNKNOWN",
-      github_id: null,
-    };
-    const project = listingProjectByKey.get(key);
-    const row: DailySeriesRow = {
-      listing_type: toListingLabel(listingType),
-      id,
-      name: meta.name,
-      author: meta.author,
-      author_alias: meta.author_alias,
-      attribution_link: meta.attribution_link,
-      project_key: project?.project_key ?? `${toListingLabel(listingType)}:${id}`,
-      project_name: project?.project_name ?? meta.name,
-      total_downloads: totalDownloads,
-    };
-    for (const snapshotDate of snapshotDates) {
-      row[snapshotDate] = dailyDeltasBySnapshot.get(`snapshot_${snapshotDate}.json`)?.get(key) ?? 0;
-    }
-    rows.push(row);
-  }
-
-  rows.sort((a, b) =>
-    Number(b.total_downloads) - Number(a.total_downloads)
-    || String(a.id).localeCompare(String(b.id)));
-  return rows;
-}
-
-function buildProjectByDayRows(
-  snapshotDates: string[],
-  latestTotals: ListingTotals,
-  dailyDeltasBySnapshot: Map<string, ListingTotals>,
-  listingMeta: Map<ListingKey, ListingMeta>,
-  listingProjectByKey: Map<ListingKey, ListingProjectRow>,
-): DailySeriesRow[] {
-  const projectRows = new Map<string, DailySeriesRow>();
-
-  for (const [key, totalDownloads] of latestTotals.entries()) {
-    const project = listingProjectByKey.get(key);
-    if (!project) continue;
-    const existing = projectRows.get(project.project_key) ?? {
-      project_key: project.project_key,
-      project_name: project.project_name,
-      author: "",
-      author_alias: "",
-      attribution_link: "",
-      listing_count: 0,
-      total_downloads: 0,
-    };
-    const meta = listingMeta.get(key) ?? {
-      name: "",
-      author: "UNKNOWN",
-      author_alias: "UNKNOWN",
-      attribution_link: "https://github.com/UNKNOWN",
-      github_id: null,
-    };
-    const authors = new Set(String(existing.author).split("; ").filter(Boolean));
-    const authorAliases = new Set(String(existing.author_alias).split("; ").filter(Boolean));
-    const attributionLinks = new Set(String(existing.attribution_link).split("; ").filter(Boolean));
-    authors.add(meta.author);
-    authorAliases.add(meta.author_alias);
-    attributionLinks.add(meta.attribution_link);
-    existing.author = [...authors].sort().join("; ");
-    existing.author_alias = [...authorAliases].sort().join("; ");
-    existing.attribution_link = [...attributionLinks].sort().join("; ");
-    existing.listing_count = Number(existing.listing_count) + 1;
-    existing.total_downloads = Number(existing.total_downloads) + totalDownloads;
-    for (const snapshotDate of snapshotDates) {
-      existing[snapshotDate] = Number(existing[snapshotDate] ?? 0)
-        + (dailyDeltasBySnapshot.get(`snapshot_${snapshotDate}.json`)?.get(key) ?? 0);
-    }
-    projectRows.set(project.project_key, existing);
-  }
-
-  return [...projectRows.values()].sort((a, b) =>
-    Number(b.total_downloads) - Number(a.total_downloads)
-    || String(a.project_key).localeCompare(String(b.project_key)));
-}
-
-function buildAuthorByDayRows(
-  snapshotDates: string[],
-  latestTotals: ListingTotals,
-  dailyDeltasBySnapshot: Map<string, ListingTotals>,
-  listingMeta: Map<ListingKey, ListingMeta>,
-): DailySeriesRow[] {
-  const authorRows = new Map<string, DailySeriesRow>();
-
-  for (const [key, totalDownloads] of latestTotals.entries()) {
-    const [listingType] = key.split(":") as ["maps" | "mods", string];
-    const meta = listingMeta.get(key) ?? {
-      name: "",
-      author: "UNKNOWN",
-      author_alias: "UNKNOWN",
-      attribution_link: "https://github.com/UNKNOWN",
-      github_id: null,
-    };
-    const existing = authorRows.get(meta.author) ?? {
-      author: meta.author,
-      author_alias: meta.author_alias,
-      attribution_link: meta.attribution_link,
-      asset_count: 0,
-      map_count: 0,
-      mod_count: 0,
-      total_downloads: 0,
-    };
-    existing.asset_count = Number(existing.asset_count) + 1;
-    if (listingType === "maps") existing.map_count = Number(existing.map_count) + 1;
-    if (listingType === "mods") existing.mod_count = Number(existing.mod_count) + 1;
-    existing.total_downloads = Number(existing.total_downloads) + totalDownloads;
-    for (const snapshotDate of snapshotDates) {
-      existing[snapshotDate] = Number(existing[snapshotDate] ?? 0)
-        + (dailyDeltasBySnapshot.get(`snapshot_${snapshotDate}.json`)?.get(key) ?? 0);
-    }
-    authorRows.set(meta.author, existing);
-  }
-
-  return [...authorRows.values()].sort((a, b) =>
-    Number(b.total_downloads) - Number(a.total_downloads)
-    || String(a.author).localeCompare(String(b.author)));
-}
-
-function buildAssetsByDayRows(
-  snapshotDates: string[],
-  dailyDeltasBySnapshot: Map<string, ListingTotals>,
-  signedDailyDeltasBySnapshot: Map<string, ListingTotals>,
-  listingIdsBySnapshot: Map<string, { maps: Set<string>; mods: Set<string> }>,
-  listingVersionsBySnapshot: Map<string, { maps: Set<string>; mods: Set<string> }>,
-): AssetByDayRow[] {
-  let cumulativeMaps = 0;
-  let cumulativeMods = 0;
-  const seenMaps = new Set<string>();
-  const seenMods = new Set<string>();
-  const seenMapVersions = new Set<string>();
-  const seenModVersions = new Set<string>();
-
-  return snapshotDates.map((snapshotDate) => {
-    let maps = 0;
-    let mods = 0;
-    let mapsSigned = 0;
-    let modsSigned = 0;
-    let newMaps = 0;
-    let newMods = 0;
-    let newMapVersions = 0;
-    let newModVersions = 0;
-    const snapshotTotals = dailyDeltasBySnapshot.get(`snapshot_${snapshotDate}.json`)
-      ?? new Map<ListingKey, number>();
-    const snapshotSignedTotals = signedDailyDeltasBySnapshot.get(`snapshot_${snapshotDate}.json`)
-      ?? new Map<ListingKey, number>();
-    const listingIds = listingIdsBySnapshot.get(`snapshot_${snapshotDate}.json`) ?? {
-      maps: new Set<string>(),
-      mods: new Set<string>(),
-    };
-    const listingVersions = listingVersionsBySnapshot.get(`snapshot_${snapshotDate}.json`) ?? {
-      maps: new Set<string>(),
-      mods: new Set<string>(),
-    };
-
-    for (const [key, totalDownloads] of snapshotTotals.entries()) {
-      const [listingType] = key.split(":") as ["maps" | "mods", string];
-      if (listingType === "maps") maps += totalDownloads;
-      if (listingType === "mods") mods += totalDownloads;
-    }
-    for (const [key, totalDownloads] of snapshotSignedTotals.entries()) {
-      const [listingType] = key.split(":") as ["maps" | "mods", string];
-      if (listingType === "maps") mapsSigned += totalDownloads;
-      if (listingType === "mods") modsSigned += totalDownloads;
-    }
-
-    for (const id of listingIds.maps) {
-      if (!seenMaps.has(id)) {
-        newMaps += 1;
-      }
-      seenMaps.add(id);
-    }
-    for (const id of listingIds.mods) {
-      if (!seenMods.has(id)) {
-        newMods += 1;
-      }
-      seenMods.add(id);
-    }
-    for (const versionKey of listingVersions.maps) {
-      if (!seenMapVersions.has(versionKey)) {
-        newMapVersions += 1;
-      }
-      seenMapVersions.add(versionKey);
-    }
-    for (const versionKey of listingVersions.mods) {
-      if (!seenModVersions.has(versionKey)) {
-        newModVersions += 1;
-      }
-      seenModVersions.add(versionKey);
-    }
-
-    cumulativeMaps += maps;
-    cumulativeMods += mods;
-
-    return {
-      snapshot_date: snapshotDate,
-      total_downloads: maps + mods,
-      maps,
-      mods,
-      total_downloads_signed: mapsSigned + modsSigned,
-      maps_signed: mapsSigned,
-      mods_signed: modsSigned,
-      total_downloads_clamped: maps + mods,
-      maps_clamped: maps,
-      mods_clamped: mods,
-      cumulative_total: cumulativeMaps + cumulativeMods,
-      cumulative_maps: cumulativeMaps,
-      cumulative_mods: cumulativeMods,
-      total_new_assets: newMaps + newMods,
-      new_maps: newMaps,
-      new_mods: newMods,
-      cumulative_assets: seenMaps.size + seenMods.size,
-      cumulative_maps_assets: seenMaps.size,
-      cumulative_mods_assets: seenMods.size,
-      total_new_assets_versions: newMapVersions + newModVersions,
-      new_maps_versions: newMapVersions,
-      new_mods_versions: newModVersions,
-      cumulative_asset_versions: seenMapVersions.size + seenModVersions.size,
-      cumulative_maps_versions: seenMapVersions.size,
-      cumulative_mods_versions: seenModVersions.size,
-    };
-  });
 }
 
 export function runGenerateAnalyticsCli(
